added update profile endpoints; fixed: delete body only with pass, remove checks if password correct; fixed: login now only with email & pass; changed paths in auth; added windows specific scripts; fixed param type in recipe;

+++ src/server/controller/auth.ts
@@ -1,6 +1,6 @@
 import { Request, Response, NextFunction } from "express";
-import { body, param } from "express-validator";
-import { Prisma, User } from "@prisma/client";
+import { body } from "express-validator";
+import { Prisma } from "@prisma/client";
 
 import prisma from "@server/common/services/prisma.service";
 import { ValidationMessages } from "@server/common/enums/validationMessages.enum";
@@ -8,9 +8,9 @@ import { ValidationMessages } from "@server/common/enums/validationMessages.enum
 /**
  * Creates a new user.
  *
- * @param User
+ * @param email
  * @param name
- * @param rating
+ * @param password
  * @return {*}
  */
 const createUser = (email: string, name: string, password: string) => {
@@ -22,25 +22,46 @@ const createUser = (email: string, name: string, password: string) => {
 };
 
 /**
- * Updates an existing user.
+ * Updates an existing user's email.
  *
- * @param name
- * @param rating
+ * @param email
  * @return {*}
  */
-const updateUser = (email?: string, name?: string, password?: string) => {
+const updateUserEmail = (email?: string) => {
   return Prisma.validator<Prisma.UserUpdateInput>()({
-    email,
-    name,
-    password,
+    email
+  });
+};
+
+/**
+ * Updates an existing user's password.
+ *
+ * @param password
+ * @return {*}
+ */
+const updateUserPassword = (password?: string) => {
+  return Prisma.validator<Prisma.UserUpdateInput>()({
+    password
+  });
+};
+
+/**
+ * Updates an existing user's name.
+ *
+ * @param name
+ * @return {*}
+ */
+const updateUserName = (name?: string) => {
+  return Prisma.validator<Prisma.UserUpdateInput>()({
+    name
   });
 };
 
 namespace UserController {
   /**
-   * Validates body payload for login and signup
+   * Validates body payload for signup
    */
-  export const validateBody = () => {
+  export const validateBodyCreate = () => {
     return [
       body("email", ValidationMessages.UNDEFINED)
         .exists()
@@ -57,8 +78,101 @@ namespace UserController {
         .isString()
         .withMessage(ValidationMessages.WRONG_TYPE)
         .isStrongPassword()
-        .withMessage(ValidationMessages.WRONG_VALUE),
+        .withMessage(ValidationMessages.WRONG_VALUE)
     ];
+  };
+
+  /**
+   * Validates body payload for login/email update
+   */
+  export const validateBody = () => {
+    return [
+      body("email", ValidationMessages.UNDEFINED)
+          .exists()
+          .isString()
+          .withMessage(ValidationMessages.WRONG_TYPE)
+          .isEmail()
+          .withMessage(ValidationMessages.WRONG_VALUE),
+      body("password", ValidationMessages.UNDEFINED)
+          .exists()
+          .isString()
+          .withMessage(ValidationMessages.WRONG_TYPE)
+          .isStrongPassword()
+          .withMessage(ValidationMessages.WRONG_VALUE)
+    ];
+  };
+
+  /**
+   * Validates body payload for password update
+   */
+  export const validateBodyUpdatePassword = () => {
+    return [
+      body("oldPassword", ValidationMessages.UNDEFINED)
+          .exists()
+          .isString()
+          .withMessage(ValidationMessages.WRONG_TYPE)
+          .isStrongPassword()
+          .withMessage(ValidationMessages.WRONG_VALUE),
+      body("newPassword", ValidationMessages.UNDEFINED)
+          .exists()
+          .isString()
+          .withMessage(ValidationMessages.WRONG_TYPE)
+          .isStrongPassword()
+          .withMessage(ValidationMessages.WRONG_VALUE)
+    ];
+  };
+
+  /**
+   * Validates body payload for name update
+   */
+  export const validateBodyUpdateName = () => {
+    return [
+      body("newName", ValidationMessages.UNDEFINED)
+          .exists()
+          .isString()
+          .withMessage(ValidationMessages.WRONG_TYPE)
+    ];
+  };
+
+  /**
+   * Validates body payload for user delete
+   */
+  export const validateBodyDelete = () => {
+    return [
+      body("password", ValidationMessages.UNDEFINED)
+          .exists()
+          .isString()
+          .withMessage(ValidationMessages.WRONG_TYPE)
+          .isStrongPassword()
+          .withMessage(ValidationMessages.WRONG_VALUE)
+    ];
+  };
+
+  /**
+   * Creates a new user
+   * @return JSON Object
+   */
+  export const create = async (
+      req: Request,
+      res: Response,
+      next: NextFunction
+  ) => {
+    try {
+      const { email, name, password } = req.body;
+
+      const user = await prisma.user.create({
+        data: createUser(email, name, password)
+      });
+
+      res.status(200).json({
+        id: user.id,
+        name: user.name,
+        email: user.email,
+        createdAt: user.createdAt,
+      });
+    } catch (err) {
+      return next(err);
+    }
   };
 
   /**
@@ -91,53 +205,100 @@ namespace UserController {
   };
 
   /**
-   * Updates the authenticated user
+   * Updates the authenticated user's email
    * @return JSON Object
    */
-  export const edit = async (
-    req: Request,
-    res: Response,
-    next: NextFunction
+  export const updateEmail = async (
+      req: Request,
+      res: Response,
+      next: NextFunction
   ) => {
     try {
       const user = req.user;
-      const { email, name, password } = req.body;
+      const { email, password } = req.body;
 
-      const recipe = await prisma.user.update({
+      const found = await prisma.user.findFirst({
         where: {
           id: user.id,
+          password: password
         },
-        data: updateUser(email, name, password),
       });
 
-      res.status(200).json(recipe);
+      if (found == null) {
+        res.status(403).json();
+      } else {
+        await prisma.user.update({
+          where: {
+            id: user.id
+          },
+          data: updateUserEmail(email)
+        });
+
+        res.status(200).json();
+      }
     } catch (err) {
       return next(err);
     }
   };
 
   /**
-   * Creates a new user
+   * Updates the authenticated user's password
    * @return JSON Object
    */
-  export const create = async (
-    req: Request,
-    res: Response,
-    next: NextFunction
+  export const updatePassword = async (
+      req: Request,
+      res: Response,
+      next: NextFunction
   ) => {
     try {
-      const { email, name, password } = req.body;
+      const user = req.user;
+      const { oldPassword, newPassword } = req.body;
 
-      const user = await prisma.user.create({
-        data: createUser(email, name, password),
+      const found = await prisma.user.findFirst({
+        where: {
+          id: user.id,
+          password: oldPassword
+        },
       });
 
-      res.status(200).json({
-        id: user.id,
-        name: user.name,
-        email: user.email,
-        createdAt: user.createdAt,
+      if (found == null) {
+        res.status(403).json();
+      } else {
+        await prisma.user.update({
+          where: {
+            id: user.id
+          },
+          data: updateUserPassword(newPassword)
+        });
+
+        res.status(200).json();
+      }
+    } catch (err) {
+      return next(err);
+    }
+  };
+
+  /**
+   * Updates the authenticated user's name
+   * @return JSON Object
+   */
+  export const updateName = async (
+      req: Request,
+      res: Response,
+      next: NextFunction
+  ) => {
+    try {
+      const user = req.user;
+      const { newName } = req.body;
+
+      await prisma.user.update({
+        where: {
+          id: user.id
+        },
+        data: updateUserName(newName)
       });
+
+      res.status(200).json();
     } catch (err) {
       return next(err);
     }
@@ -154,16 +315,28 @@ namespace UserController {
   ) => {
     try {
       const user = req.user;
+      const { password } = req.body;
 
-      const removed = await prisma.user.delete({
-        where: { id: user.id },
+      const found = await prisma.user.findFirst({
+        where: {
+          id: user.id,
+          password: password
+        },
       });
 
-      //TODO: CASCADE down
+      if (found == null) {
+        res.status(403).json();
+      } else {
+        await prisma.user.delete({
+          where: {
+            id: user.id
+          }
+        });
 
-      res.status(200).json({
-        id: removed.id,
-      });
+        //TODO: CASCADE down
+
+        res.status(200).json();
+      }
     } catch (err) {
       return next(err);
     }
